requirements.txt
Django~=4.2.7 
django-seed~=0.3.1
psycopg2-binary~=2.9.9
django-bootstrap-v5~=1.0.11
django-bootstrap-icons~=0.8.6
djangorestframework~=3.14.0